Includes #36-Replace "float" with "double, #35-"Add underground cave system", #34-"Increase playability of the terrain", #33-"Allow chunk generationin the area "x<0" and "z<0"

+++ src/com/github/begla/blockmania/datastructures/BlockPosition.java
@@ -39,13 +39,13 @@ public final class BlockPosition implements Comparable<BlockPosition> {
         this._origin = origin;
     }
 
-    float getDistance() {
-        return VectorPool.getVector((float) x - _origin.x, (float) y - _origin.y, (float) z - _origin.z).length();
+    double getDistance() {
+        return VectorPool.getVector(x - _origin.x, y - _origin.y, z - _origin.z).length();
     }
 
     public int compareTo(BlockPosition o) {
-        float distance = getDistance();
-        float oDistance = o.getDistance();
+        double distance = getDistance();
+        double oDistance = o.getDistance();
 
         if (oDistance > distance)
             return -1;
